+++ src/App.js
@@ -5,14 +5,15 @@ import React, { useState } from "react";
         functional component. 'state' enables dynamic webs for data that changes,
         else, it would be a static page
     */
-import Navigation from './components/common/Navigation'
-import './App.css'
-import './styles/index.css'
-import Sleep from './components/Sleep'
-import Fitness from './components/Fitness'
-import Hobbies from './components/Hobbies'
-import School from './components/School'
-import Chores from './components/Chores'
+import Navigation from './components/common/Navigation';
+import './App.css';
+import './styles/index.css';
+import Sleep from './components/Sleep/Sleep';
+import Fitness from './components/Fitness/Fitness';
+import Hobbies from './components/Hobbies/Hobbies';
+import School from './components/School/School';
+import Chores from './components/Chores/Chores';
+import Weight from './components/Weight/Weight';
 
 function App() {
     const [currentView, setCurrentView] = useState('sleep');