+++ CHANGELOG.md
@@ -38,6 +38,14 @@ Track:
 - App.js
 - utils folder
     - storage.js
+- New Branch (Dev)
+    - feature/sleep-tracker
+        - sleepCalculations.js
+            - calculate sleep hours
+            - get sleep in range
+            - calc median sleep hours
+            - calc median bed time
+            - calc median wake time
 
 ### Removed
 - App.test